+++ server/src/providers/jwt/interfaces/ITokenProperties.ts
