+++ server/src/routes/account/account.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { AccountService } from './account.service';
 import { AccountController } from './account.controller';
-import { JwtModule } from '../../services/jwt/jwt.module';
+import { JwtModule } from '../../providers/jwt/jwt.module';
 import { UserModule } from '../../database/models/user/user.module';
 
 @Module({
